requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
decorator==5.2.1
distlib==0.3.9
Django==5.2
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
executing==2.2.0
filelock==3.18.0
flake8==7.2.0
gunicorn==23.0.0
identify==2.6.9
idna==3.10
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.2.3
virtualenv==20.30.0
wcwidth==0.2.13
whitenoise==6.9.0
